Changed render on page to router. Added logout on calendar. Reading userId from local storage on Calendar
