+++ calendr_django/calendr-vue-app/src/App.vue
@@ -1,66 +1,19 @@
 <template>
 	<div id="app">
-		<!-- <router-view header=""></router-view> -->
+		<router-view header=""></router-view>
 		<!-- example: <router-link to="/" name="home">Home</router-link> -->
 
-		<Home
-			v-if="!user"
-			@handleSignup="handleSignup"
-			@handleLogin="handleLogin"
-			:signin="signin"
-			:login="login"
-			:user="user"
-			@fieldChange="fieldChange"
-		/>
-
-		<div v-else>
-			<header>
-				<button @click="clearUser">Logout</button>
-			</header>
-			<Calendar :user="user" />
-		</div>
 	</div>
 </template>
 
 <script>
-import Calendar from './components/Calendar.vue';
-import Home from './components/Home.vue';
-import { CreateUser, FindUsername } from './services/endpoints';
+// import Calendar from './components/Calendar.vue';
+// import Home from './components/Home.vue';
 export default {
 	name: 'App',
-	data: () => ({
-		signin: '',
-		login: '',
-		user: JSON.parse(localStorage.getItem('user')) || null,
-	}),
-	components: {
-		Home,
-		Calendar,
-	},
-	methods: {
-		fieldChange(name, value) {
-			this[name] = value;
-		},
 
-		async handleSignup() {
-			console.log(this.signin);
-			console.log(`${this.signin} signed up`);
-			const user = await CreateUser(this.signin);
-			localStorage.setItem('user', JSON.stringify(user));
-		},
-		async handleLogin() {
-			console.log(this.login);
-			const user = await FindUsername(this.login);
-			localStorage.setItem('user', JSON.stringify(user));
-			this.user = user;
-			console.log(user);
-			console.log(`${this.login} logged in`);
-		},
-		clearUser() {
-			localStorage.clear();
-			this.user = null;
-			// this.username = ''
-		},
+	methods: {
+
 	},
 };
 </script>
